requirements.txt
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fastapi==0.108.0
h11==0.14.0
idna==3.6
ipykernel==6.27.1
ipython==8.18.1
iso8601==1.1.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
platformdirs==4.1.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.32.0.post1
tornado==6.4
tortoise-orm==0.20.0
traitlets==5.14.0
typing_extensions==4.9.0
uvicorn==0.25.0
wcwidth==0.2.12
